Reroll and Weight Environments
The StageBlocker, when the player cannot move to the next stage, will
adjust the environment selection. First occurance rerolls the
environment within the same stage, second occurance rerolls back to stage 1.

The LocationHandler will add weight to environments based on the number
of remaining locations. This should help the player not get stuck where
RNG is taking the player to environments which do not have locations.