Count all chests and shrines
When in environments that do not have remaining locations (or
environments that never had locations), count the chests and shrines
used but don't let them complete locations. This is so the player can
work towards locations even when they are somewhere they cannot finish
the location, for the sake of fairness.